requirements.txt
fastapi
dotenv
uvicorn
pydantic-settings
pyyaml